The program can now find your "author name" from your GitHub username
This is way harder than is sounds and I'm super happy that I managed to
do this! When making a commit, the actual commit data will that you
access using libgit2, will list the "author name". This is the name that
you set in Git!

This means that this will be different from your GitHub username.

Using some web scraping in combination with GitHubs API, I was able to
search a repository and find the "author name" for a given GitHub user!
YAY! :D